Move facade package into models package
