+++ src/main/java/org/fundacion/jala/converter/controller/ExtractMetadataController.java
@@ -10,16 +10,12 @@ package org.fundacion.jala.converter.controller;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.fundacion.jala.converter.facade.ExtractFacade;
-import org.fundacion.jala.converter.service.ExtractMetadata;
+import org.fundacion.jala.converter.models.facade.ExtractFacade;
 import org.fundacion.jala.converter.service.FileStorageService;
-import org.fundacion.jala.converter.service.ObjectMetadata;
-import org.fundacion.jala.converter.service.metadata.TypeFileExport;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import java.io.File;
 import java.io.IOException;
 
 @RestController
