+++ src/main/java/org/fundacion/jala/converter/controller/JavaCompilerController.java
@@ -10,8 +10,7 @@ package org.fundacion.jala.converter.controller;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.fundacion.jala.converter.facade.CompilerFacade;
-import org.fundacion.jala.converter.service.javacompiler.JavaCompiler;
+import org.fundacion.jala.converter.models.facade.CompilerFacade;
 import org.fundacion.jala.converter.service.javacompiler.JavaVersion;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
