+++ src/main/java/org/fundacion/jala/converter/controller/PythonCompilerController.java
@@ -10,13 +10,10 @@ package org.fundacion.jala.converter.controller;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.fundacion.jala.converter.facade.CompilerFacade;
-import org.fundacion.jala.converter.service.FileStorageService;
+import org.fundacion.jala.converter.models.facade.CompilerFacade;
 import org.fundacion.jala.converter.service.pythoncompiler.Python;
-import org.fundacion.jala.converter.service.pythoncompiler.*;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
+
 import java.io.IOException;
 
 @RestController
